+++ app/controllers/static_invites_controller.rb
@@ -1,8 +1,8 @@
 class StaticInvitesController < ApplicationController
-  before_action :find_character, only: %i[new]
+  before_action :find_characters, only: %i[find]
   before_action :find_static_invite, only: %i[approve decline]
 
-  def new; end
+  def find; end
 
   def approve
     ApproveStaticInvite.call(static_invite: @static_invite, static: @static_invite.static, character: @static_invite.character, status: 2)
@@ -16,9 +16,8 @@ class StaticInvitesController < ApplicationController
 
   private
 
-  def find_character
-    @character = Character.where(user_id: Current.user.id).find_by(id: params[:character_id])
-    render_error(t('custom_errors.object_not_found'), 404) if @character.nil?
+  def find_characters
+    @user_characters = ActiveModelSerializers::SerializableResource.new(Character.where(user: Current.user).includes(race: :fraction), each_serializer: CharacterIndexSerializer).as_json[:characters]
   end
 
   def find_static_invite