+++ app/policies/static_invite_policy.rb
@@ -6,4 +6,9 @@ class StaticInvitePolicy < ApplicationPolicy
     return user.any_static_role?(static) if record == 'true'
     character.user_id == user.id
   end
+
+  def approve?
+    return character.user_id == user.id if record == 'true'
+    user.any_static_role?(static)
+  end
 end