+++ spec/factories/static_invites.rb
@@ -1,5 +1,6 @@
 FactoryBot.define do
   factory :static_invite do
+    from_static { true }
     association :static, :guild
     association :character
   end